CustomLists Views cleanup
